make REFTYPE_CHOICES common
