+++ src/model/Model_Attachment.h
@@ -27,29 +27,6 @@ class Model_Attachment : public Model<DB_Table_ATTACHMENT_V1>
 {
 public:
     using Model<DB_Table_ATTACHMENT_V1>::get;
-    enum REFTYPE_ID {
-        REFTYPE_ID_TRANSACTION = 0,
-        REFTYPE_ID_STOCK,
-        REFTYPE_ID_ASSET,
-        REFTYPE_ID_BANKACCOUNT,
-        REFTYPE_ID_BILLSDEPOSIT,
-        REFTYPE_ID_PAYEE,
-        REFTYPE_ID_TRANSACTIONSPLIT,
-        REFTYPE_ID_BILLSDEPOSITSPLIT,
-        REFTYPE_ID_size
-    };
-    static const wxString REFTYPE_NAME_TRANSACTION;
-    static const wxString REFTYPE_NAME_STOCK;
-    static const wxString REFTYPE_NAME_ASSET;
-    static const wxString REFTYPE_NAME_BANKACCOUNT;
-    static const wxString REFTYPE_NAME_BILLSDEPOSIT;
-    static const wxString REFTYPE_NAME_PAYEE;
-    static const wxString REFTYPE_NAME_TRANSACTIONSPLIT;
-    static const wxString REFTYPE_NAME_BILLSDEPOSITSPLIT;
-
-    static ChoicesName REFTYPE_CHOICES;
-    static const wxString reftype_name(int id);
-    static int reftype_id(const wxString& name, int default_id = -1);
 
 public:
     Model_Attachment();
@@ -86,17 +63,5 @@ public:
     /** Return all attachments descriptions*/
     wxArrayString allDescriptions();
 };
-
-//----------------------------------------------------------------------------
-
-inline const wxString Model_Attachment::reftype_name(int id)
-{
-    return REFTYPE_CHOICES.getName(id);
-}
-
-inline int Model_Attachment::reftype_id(const wxString& name, int default_id)
-{
-    return REFTYPE_CHOICES.findName(name, default_id);
-}
 
 #endif
